pyproject.toml
[project]
name = "icebreaker-mcp"
version = "0.1.0"
description = "Intelligent Snowflake MCP Agent specialized in DataOps and Administration"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Icebreaker Team" },
]
keywords = ["mcp", "snowflake", "dataops", "administration", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "fastmcp>=2.12.4",
    "snowflake-connector-python>=3.15.0",
    "snowflake-core>=1.0.0",
    "sqlglot>=27.8.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "structlog>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
icebreaker-mcp = "icebreaker.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/icebreaker"]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "I", "UP", "B", "A", "C4", "T20"]
ignore = ["E501", "B008"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-q -m 'not integration'"
markers = [
    "unit: Fast tests without external dependencies (default)",
    "integration: Tests requiring live Snowflake connection",
    "contract: MCP tool contract verification tests",
    "safety: Safety mechanism and security tests",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.4",
]
